Added support for EventName to not require a parameter.
If not specified, the `Code` property will be set the the class the attribute is decorating.

I wrote tests for most of the TypeFinder class. I didn't see a way to register a duplicate event easily as the search is done in the constructor of the class and would fail all tests 100% of the time due to the exception.

I think future work has to refactor that class to make it more testable and modular.